+++ libcnb/src/detect.rs
@@ -1,7 +1,7 @@
 //! Provides detect phase specific types and helpers.
 
 use crate::buildpack::Buildpack;
-use crate::data::buildpack::StackId;
+use crate::target::ContextTarget;
 use crate::{data::build_plan::BuildPlan, data::buildpack::ComponentBuildpackDescriptor};
 use std::fmt::Debug;
 use std::path::PathBuf;
@@ -10,7 +10,7 @@ use std::path::PathBuf;
 pub struct DetectContext<B: Buildpack + ?Sized> {
     pub app_dir: PathBuf,
     pub buildpack_dir: PathBuf,
-    pub stack_id: StackId,
+    pub target: ContextTarget,
     pub platform: B::Platform,
     pub buildpack_descriptor: ComponentBuildpackDescriptor<B::Metadata>,
 }