+++ libcnb/src/runtime.rs
@@ -1,10 +1,11 @@
 use crate::build::{BuildContext, InnerBuildResult};
 use crate::buildpack::Buildpack;
-use crate::data::buildpack::{BuildpackApi, StackId};
+use crate::data::buildpack::BuildpackApi;
 use crate::detect::{DetectContext, InnerDetectResult};
 use crate::error::Error;
 use crate::platform::Platform;
 use crate::sbom::cnb_sbom_path;
+use crate::target::ContextTarget;
 #[cfg(feature = "trace")]
 use crate::tracing::start_trace;
 use crate::util::is_not_found_error_kind;
@@ -132,38 +133,33 @@ pub fn libcnb_runtime_detect<B: Buildpack>(
     #[cfg(feature = "trace")]
     let mut trace = start_trace(&buildpack_descriptor.buildpack, "detect");
 
+    #[cfg(feature = "trace")]
     let mut trace_error = |err: &dyn std::error::Error| {
-        #[cfg(feature = "trace")]
         trace.set_error(err);
     };
-    let stack_id: StackId = env::var("CNB_STACK_ID")
-        .map_err(Error::CannotDetermineStackId)
-        .and_then(|stack_id_string| stack_id_string.parse().map_err(Error::StackIdError))
-        .map_err(|err| {
-            trace_error(&err);
-            err
-        })?;
 
-    let platform = B::Platform::from_path(&args.platform_dir_path).map_err(|inner_err| {
-        let err = Error::CannotCreatePlatformFromPath(inner_err);
-        trace_error(&err);
-        err
-    })?;
+    #[cfg(not(feature = "trace"))]
+    let mut trace_error = |_: &dyn std::error::Error| {};
+
+    let platform = B::Platform::from_path(&args.platform_dir_path)
+        .map_err(Error::CannotCreatePlatformFromPath)
+        .inspect_err(|err| trace_error(err))?;
 
     let build_plan_path = args.build_plan_path;
+
+    let target = context_target().inspect_err(|err| trace_error(err))?;
 
     let detect_context = DetectContext {
         app_dir,
         buildpack_dir,
-        stack_id,
+        target,
         platform,
         buildpack_descriptor,
     };
 
-    let detect_result = buildpack.detect(detect_context).map_err(|err| {
-        trace_error(&err);
-        err
-    })?;
+    let detect_result = buildpack
+        .detect(detect_context)
+        .inspect_err(|err| trace_error(err))?;
 
     match detect_result.0 {
         InnerDetectResult::Fail => {
@@ -173,11 +169,9 @@ pub fn libcnb_runtime_detect<B: Buildpack>(
         }
         InnerDetectResult::Pass { build_plan } => {
             if let Some(build_plan) = build_plan {
-                write_toml_file(&build_plan, build_plan_path).map_err(|inner_err| {
-                    let err = Error::CannotWriteBuildPlan(inner_err);
-                    trace_error(&err);
-                    err
-                })?;
+                write_toml_file(&build_plan, build_plan_path)
+                    .map_err(Error::CannotWriteBuildPlan)
+                    .inspect_err(|err| trace_error(err))?;
             }
             #[cfg(feature = "trace")]
             trace.add_event("detect-passed");
@@ -190,6 +184,7 @@ pub fn libcnb_runtime_detect<B: Buildpack>(
 ///
 /// Exposed only to allow for advanced use-cases where build is programmatically invoked.
 #[doc(hidden)]
+#[allow(clippy::too_many_lines)]
 pub fn libcnb_runtime_build<B: Buildpack>(
     buildpack: &B,
     args: BuildArgs,
@@ -206,56 +201,45 @@ pub fn libcnb_runtime_build<B: Buildpack>(
     #[cfg(feature = "trace")]
     let mut trace = start_trace(&buildpack_descriptor.buildpack, "build");
 
+    #[cfg(feature = "trace")]
     let mut trace_error = |err: &dyn std::error::Error| {
-        #[cfg(feature = "trace")]
         trace.set_error(err);
     };
 
-    let stack_id: StackId = env::var("CNB_STACK_ID")
-        .map_err(Error::CannotDetermineStackId)
-        .and_then(|stack_id_string| stack_id_string.parse().map_err(Error::StackIdError))
-        .map_err(|err| {
-            trace_error(&err);
-            err
-        })?;
+    #[cfg(not(feature = "trace"))]
+    let mut trace_error = |_: &dyn std::error::Error| {};
 
-    let platform = Platform::from_path(&args.platform_dir_path).map_err(|inner_err| {
-        let err = Error::CannotCreatePlatformFromPath(inner_err);
-        trace_error(&err);
-        err
-    })?;
+    let platform = Platform::from_path(&args.platform_dir_path)
+        .map_err(Error::CannotCreatePlatformFromPath)
+        .inspect_err(|err| trace_error(err))?;
 
-    let buildpack_plan = read_toml_file(&args.buildpack_plan_path).map_err(|inner_err| {
-        let err = Error::CannotReadBuildpackPlan(inner_err);
-        trace_error(&err);
-        err
-    })?;
+    let buildpack_plan = read_toml_file(&args.buildpack_plan_path)
+        .map_err(Error::CannotReadBuildpackPlan)
+        .inspect_err(|err| trace_error(err))?;
 
     let store = match read_toml_file::<Store>(layers_dir.join("store.toml")) {
         Err(TomlFileError::IoError(io_error)) if is_not_found_error_kind(&io_error) => Ok(None),
         other => other.map(Some),
     }
     .map_err(Error::CannotReadStore)
-    .map_err(|err| {
-        trace_error(&err);
-        err
-    })?;
+    .inspect_err(|err| trace_error(err))?;
+
+    let target = context_target().inspect_err(|err| trace_error(err))?;
 
     let build_context = BuildContext {
         layers_dir: layers_dir.clone(),
         app_dir,
-        stack_id,
         platform,
+        target,
         buildpack_plan,
         buildpack_dir,
         buildpack_descriptor,
         store,
     };
 
-    let build_result = buildpack.build(build_context).map_err(|err| {
-        trace_error(&err);
-        err
-    })?;
+    let build_result = buildpack
+        .build(build_context)
+        .inspect_err(|err| trace_error(err))?;
 
     match build_result.0 {
         InnerBuildResult::Pass {
@@ -265,19 +249,15 @@ pub fn libcnb_runtime_build<B: Buildpack>(
             launch_sboms,
         } => {
             if let Some(launch) = launch {
-                write_toml_file(&launch, layers_dir.join("launch.toml")).map_err(|inner_err| {
-                    let err = Error::CannotWriteLaunch(inner_err);
-                    trace_error(&err);
-                    err
-                })?;
+                write_toml_file(&launch, layers_dir.join("launch.toml"))
+                    .map_err(Error::CannotWriteLaunch)
+                    .inspect_err(|err| trace_error(err))?;
             };
 
             if let Some(store) = store {
-                write_toml_file(&store, layers_dir.join("store.toml")).map_err(|inner_err| {
-                    let err = Error::CannotWriteStore(inner_err);
-                    trace_error(&err);
-                    err
-                })?;
+                write_toml_file(&store, layers_dir.join("store.toml"))
+                    .map_err(Error::CannotWriteStore)
+                    .inspect_err(|err| trace_error(err))?;
             };
 
             for build_sbom in build_sboms {
@@ -286,10 +266,7 @@ pub fn libcnb_runtime_build<B: Buildpack>(
                     &build_sbom.data,
                 )
                 .map_err(Error::CannotWriteBuildSbom)
-                .map_err(|err| {
-                    trace_error(&err);
-                    err
-                })?;
+                .inspect_err(|err| trace_error(err))?;
             }
 
             for launch_sbom in launch_sboms {
@@ -298,10 +275,7 @@ pub fn libcnb_runtime_build<B: Buildpack>(
                     &launch_sbom.data,
                 )
                 .map_err(Error::CannotWriteLaunchSbom)
-                .map_err(|err| {
-                    trace_error(&err);
-                    err
-                })?;
+                .inspect_err(|err| trace_error(err))?;
             }
 
             #[cfg(feature = "trace")]
@@ -381,5 +355,24 @@ fn read_buildpack_descriptor<BD: DeserializeOwned, E: Debug>() -> crate::Result<
     read_buildpack_dir().and_then(|buildpack_dir| {
         read_toml_file(buildpack_dir.join("buildpack.toml"))
             .map_err(Error::CannotReadBuildpackDescriptor)
+    })
+}
+
+fn context_target<E>() -> crate::Result<ContextTarget, E>
+where
+    E: Debug,
+{
+    let os = env::var("CNB_TARGET_OS").map_err(Error::CannotDetermineTargetOs)?;
+    let arch = env::var("CNB_TARGET_ARCH").map_err(Error::CannotDetermineTargetArch)?;
+    let arch_variant = env::var("CNB_TARGET_ARCH_VARIANT").ok();
+    let distro_name = env::var("CNB_TARGET_DISTRO_NAME").ok();
+    let distro_version = env::var("CNB_TARGET_DISTRO_VERSION").ok();
+
+    Ok(ContextTarget {
+        os,
+        arch,
+        arch_variant,
+        distro_name,
+        distro_version,
     })
 }